Add caching of partial indicies.
This commit adds the capability to compare the edit time of partial
indice files and their corresponding source files and only recreate
the partial indices if the source file has been altered since the
partial indices were created.

Moreover, it creates a memory of what source files were examined
in the previous execution. If in the current execution the a
source file froma previous execution is no longer included,
any generated partial indices for this file are purged. This
effectively prevents false positives from past searches.